Remove infinite loop when js ends with comment
Fixes Issue #1
To reproduce: jsmin.jsmin("//test\r\n") This will hang indefinately.

We remove the empty string as a valid newline character when eating up
comments. The empty string is namely used as EOF marker and the code should
stop when it reads ''.

We also added a unittest for this case.